HW-3 Домашка №3
