+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/config/Config.java
@@ -11,4 +11,7 @@ public interface Config {
   String spendUrl();
 
   String ghUrl();
+  String auth();
+  String gateway();
+  String userdataUrl();
 }
