+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/meta/User.java
@@ -1,0 +1,25 @@
+package guru.qa.niffler.jupiter.annotation.meta;
+
+import guru.qa.niffler.jupiter.annotation.Category;
+import guru.qa.niffler.jupiter.annotation.Spending;
+import guru.qa.niffler.jupiter.extension.CategoryExtension;
+import guru.qa.niffler.jupiter.extension.SpendingExtension;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+@ExtendWith({
+    CategoryExtension.class,
+    SpendingExtension.class
+})
+public @interface User {
+
+    Category[] categories() default {};
+
+    Spending[] spendings() default {};
+}
