+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/SpendingExtension.java
@@ -2,6 +2,7 @@ package guru.qa.niffler.jupiter.extension;
 
 import guru.qa.niffler.api.SpendApiClient;
 import guru.qa.niffler.jupiter.annotation.Spending;
+import guru.qa.niffler.jupiter.annotation.meta.User;
 import guru.qa.niffler.model.CategoryJson;
 import guru.qa.niffler.model.CurrencyValues;
 import guru.qa.niffler.model.SpendJson;
@@ -22,26 +23,30 @@ public class SpendingExtension implements BeforeEachCallback, ParameterResolver 
 
   @Override
   public void beforeEach(ExtensionContext context) throws Exception {
-    AnnotationSupport.findAnnotation(context.getRequiredTestMethod(), Spending.class)
+    AnnotationSupport.findAnnotation(context.getRequiredTestMethod(), User.class)
         .ifPresent(anno -> {
-          SpendJson spend = new SpendJson(
-              null,
-              new Date(),
-              new CategoryJson(
-                  null,
-                  anno.category(),
-                  anno.username(),
-                  false
-              ),
-              CurrencyValues.RUB,
-              anno.amount(),
-              anno.description(),
-              anno.username()
-          );
-          context.getStore(NAMESPACE).put(
-              context.getUniqueId(),
-              spendApiClient.createSpend(spend)
-          );
+            if(anno.spendings().length > 0){
+                Spending annoSpending = anno.spendings()[0];
+                SpendJson spend = new SpendJson(
+                    null,
+                    new Date(),
+                    new CategoryJson(
+                        null,
+                        annoSpending.category(),
+                        annoSpending.username(),
+                        false
+                    ),
+                    CurrencyValues.RUB,
+                    annoSpending.amount(),
+                    annoSpending.description(),
+                    annoSpending.username()
+                );
+                context.getStore(NAMESPACE).put(
+                    context.getUniqueId(),
+                    spendApiClient.createSpend(spend)
+                );
+            }
+
         });
   }
