+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/LogInTest.java
@@ -5,7 +5,7 @@ import guru.qa.niffler.config.Config;
 import guru.qa.niffler.page.LoginPage;
 import org.junit.jupiter.api.Test;
 
-public class LogInWebTest {
+public class LogInTest {
 
     private static final Config CFG = Config.getInstance();
     @Test
