+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/ProfileTest.java
@@ -3,42 +3,66 @@ package guru.qa.niffler.test.web;
 import com.codeborne.selenide.Selenide;
 import guru.qa.niffler.config.Config;
 import guru.qa.niffler.jupiter.annotation.Category;
-import guru.qa.niffler.jupiter.annotation.meta.WebTest;
+import guru.qa.niffler.jupiter.annotation.meta.User;
+import guru.qa.niffler.jupiter.extension.BrowserExtension;
 import guru.qa.niffler.model.CategoryJson;
 import guru.qa.niffler.page.LoginPage;
-import guru.qa.niffler.page.ProfilePage;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-@WebTest
+@ExtendWith(BrowserExtension.class)
 public class ProfileTest {
 
-  private static final Config CFG = Config.getInstance();
+    private static final Config CFG = Config.getInstance();
 
-  @Category(
-      username = "duck",
-      archived = true
-  )
-  @Test
-  void archivedCategoryShouldPresentInCategoriesList(CategoryJson category) {
-    Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .successLogin("duck", "12345")
-        .checkThatPageLoaded();
 
-    Selenide.open(CFG.frontUrl() + "profile", ProfilePage.class)
-        .checkArchivedCategoryExists(category.name());
-  }
+    @User(
+        categories = @Category(
+            username = "PetyaMain",
+            name = "archivedCategory",
+            archived = true
+        )
+    )
+    @Test
+    public void archivedCategoryShouldPresentInCategoriesList(CategoryJson category){
 
-  @Category(
-      username = "duck",
-      archived = false
-  )
-  @Test
-  void activeCategoryShouldPresentInCategoriesList(CategoryJson category) {
-    Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .successLogin("duck", "12345")
-        .checkThatPageLoaded();
+        Selenide.open(CFG.frontUrl(), LoginPage.class)
+            .login("PetyaMain", "123")
+            .clickToProfile()
+            .clickToShowArchivedToggle()
+            .searchingCategoryShouldBeExist(category.name());
+    }
 
-    Selenide.open(CFG.frontUrl() + "profile", ProfilePage.class)
-        .checkCategoryExists(category.name());
-  }
+    @User(
+        categories = @Category(
+            username = "PetyaMain",
+            name = "activeCategory",
+            archived = false
+        )
+    )
+    @Test
+    public void activeCategoryShouldPresentInCategoriesList(CategoryJson category){
+
+        Selenide.open(CFG.frontUrl(), LoginPage.class)
+            .login("PetyaMain", "123")
+            .clickToProfile()
+            .searchingCategoryShouldBeExist(category.name());
+    }
+
+    @User(
+        categories = @Category(
+            username = "PetyaMain",
+            name = "archivedCategoryNotPresent",
+            archived = true
+        )
+    )
+    @Test
+    public void archivedCategoryShouldNotPresentInCategoriesList(CategoryJson category){
+
+        Selenide.open(CFG.frontUrl(), LoginPage.class)
+            .login("PetyaMain", "123")
+            .clickToProfile()
+            .searchingCategoryShouldNotBeExist(category.name());
+    }
+
 }
