+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/utils/RandomDataUtils.java
@@ -1,0 +1,24 @@
+package guru.qa.niffler.utils;
+
+import com.github.javafaker.Faker;
+
+public class RandomDataUtils {
+    private static final Faker faker = new Faker();
+
+    public static String randomUserName(){
+        return  faker.name().username();
+    }
+    public static String randomName(){
+        return faker.name().name();
+    }
+    public static String randomSurname(){
+        return faker.name().lastName();
+    }
+    public static String randomCategoryName(){
+        return faker.random().hex();
+    }
+    public static String randomPassword(){
+        return faker.internet().password();
+    }
+
+}
